+++ flash_attn/flash_attn_triton_amd/fwd_prefill.py
@@ -1,7 +1,7 @@
 import torch
 import triton
 import triton.language as tl
-from .common import get_shape_from_layout, get_strides_from_layout
+from .utils import get_shape_from_layout, get_strides_from_layout
 
 DEBUG = False
